fix #3

+++ function.cpp
@@ -1274,7 +1274,7 @@ void	CFunction::ExecHistoryP2(CCell& cell, CStatement &st)
 	if (index < 0)
 		return;
 
-	ptrdiff_t start = std::min(aya::int_t(st.cell().size())-1,cell.ansv_const().GetValueInt());
+	ptrdiff_t start = (ptrdiff_t)std::min(aya::int_t(st.cell().size())-1,cell.ansv_const().GetValueInt());
 
 	for(ptrdiff_t i = start ; i >= 0; i--) {
 		if (st.cell()[i].value_GetType() == F_TAG_STRING_EMBED) {
